pyproject.toml
[project]
name = "ms-boilerplate"
version = "0.1.0"
description = "A boilerplate to develop microservices in FastAPI"
authors = [
    {name = "Prashant",email = "officialprashant110@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "attrs (>=25.1.0,<26.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "boto3 (>=1.37.9,<2.0.0)",
    "ddtrace (>=3.2.1,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "structlog (>=25.1.0,<26.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pysqslistener (>=0.9.0,<0.10.0)",
    "mypy-boto3-s3 (>=1.37.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
